wallet: move can_spend to wallet.c and json_addfunds to walletrpc.c
